+++ src/main/java/nu/revitalized/revitalizedwebshop/services/GarmentService.java
@@ -98,8 +98,8 @@ public class GarmentService {
             Double maxPrice,
             Integer minStock,
             Integer maxStock,
-            Double minRating,
-            Double maxRating,
+            Integer minRating,
+            Integer maxRating,
             String size,
             String color
     ) {
